wip on escrow sweep

+++ contracts/test/ValueTierTest.sol
@@ -11,7 +11,7 @@ import { Tier, ITier } from "../tier/ITier.sol";
 /// testing of `internal` functions.
 contract ValueTierTest is ValueTier {
     /// Set the `tierValues` on construction to be referenced immutably.
-    constructor(uint256[8] memory tierValues_)
+    constructor(uint[8] memory tierValues_)
         ValueTier(tierValues_)
     { } // solhint-disable-line no-empty-blocks
 
@@ -19,13 +19,13 @@ contract ValueTierTest is ValueTier {
     function wrappedTierToValue(Tier tier_)
         external
         view
-        returns(uint256)
+        returns(uint)
     {
         return ValueTier.tierToValue(tier_);
     }
 
     /// Wraps `valueToTier`.
-    function wrappedValueToTier(uint256 value_)
+    function wrappedValueToTier(uint value_)
         external
         view
         returns(Tier)
